requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
click==8.2.0
discord.py==2.5.2
ffmpeg-python==0.2.0
Flask==3.1.1
frozenlist==1.6.0
future==1.0.0
idna==3.10
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
propcache==0.3.1
pycord==0.1.1
python-dotenv==1.1.0
Werkzeug==3.1.3
yarl==1.20.0
flask
